Introduced localization for QueryBuilder DateField and updated example
