+++ web/client/examples/queryform/app.jsx
@@ -1,33 +1,22 @@
+/**
+ * Copyright 2016, GeoSolutions Sas.
+ * All rights reserved.
+ *
+ * This source code is licensed under the BSD-style license found in the
+ * LICENSE file in the root directory of this source tree.
+ */
 const React = require('react');
 const ReactDOM = require('react-dom');
 
-const {connect, Provider} = require('react-redux');
-
-// include application component
-const QueryBuilder = require('../../components/QueryForm/QueryBuilder');
+const {Provider} = require('react-redux');
 
 // initializes Redux store
 var store = require('./stores/queryformstore');
 
-const {bindActionCreators} = require('redux');
-const {addFilterField, removeFilterField,
-    updateFilterField, updateExceptionField} = require('../../actions/queryform');
+const {loadLocale} = require('../../actions/locale');
+const LocaleUtils = require('../../utils/LocaleUtils');
 
-// connecting a Dumb component to the store
-// makes it a smart component
-// we both connect state => props
-// and actions to event handlers
-const SmartQueryForm = connect((state) => {
-    return {
-        filterFields: state.queryForm.filterFields,
-        attributes: state.queryForm.attributes
-    };
-}, (dispatch) => bindActionCreators({
-    onAddFilterField: addFilterField,
-    onRemoveFilterField: removeFilterField,
-    onUpdateFilterField: updateFilterField,
-    onUpdateExceptionField: updateExceptionField
-}, dispatch))(QueryBuilder);
+const QueryForm = require('./containers/QueryForm');
 
 // we spread the store to the all application
 // wrapping it with a Provider component
@@ -35,10 +24,26 @@ const QueryFormApp = React.createClass({
     render() {
         return (
         <Provider store={store}>
-            <SmartQueryForm/>
+            <QueryForm/>
         </Provider>);
     }
 });
 
-// Renders the application, wrapped by the Redux Provider to connect the store to components
-ReactDOM.render(<QueryFormApp/>, document.getElementById('container'));
+let locale = LocaleUtils.getUserLocale();
+store.dispatch(loadLocale('../../translations', locale));
+
+const startApp = () => {
+    // Renders the application, wrapped by the Redux Provider to connect the store to components
+    ReactDOM.render(<QueryFormApp/>, document.getElementById('container'));
+};
+
+if (!global.Intl ) {
+    require.ensure(['intl', 'intl/locale-data/jsonp/en.js', 'intl/locale-data/jsonp/it.js'], (require) => {
+        global.Intl = require('intl');
+        require('intl/locale-data/jsonp/en.js');
+        require('intl/locale-data/jsonp/it.js');
+        startApp();
+    });
+} else {
+    startApp();
+}
