+++ web/client/examples/queryform/stores/queryformstore.js
@@ -1,6 +1,9 @@
-var {createStore, combineReducers} = require('redux');
+var {combineReducers} = require('redux');
+
+var DebugUtils = require('../../../utils/DebugUtils');
 
 var queryForm = require('../../../reducers/queryform');
+var locale = require('../../../reducers/locale');
 
 const initialState = {
     filterFields: [
@@ -33,8 +36,9 @@ const initialState = {
 
  // reducers
 const reducers = combineReducers({
-    queryForm
+    queryForm,
+    locale
 });
 
 // export the store with the given reducers
-module.exports = createStore(reducers, {queryForm: initialState});
+module.exports = DebugUtils.createDebugStore(reducers, {queryForm: initialState});
